requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.5.0
pathspec==0.10.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.3.0
Unipath==1.1
